Moved tests into a lib dir
This will make room for command tests
