+++ test/fixtures/loader/duplicates.yaml
