+++ test/fixtures/loader/json-schema/a.json
