+++ test/fixtures/loader/json-schema/openapi.yaml
