+++ test/fixtures/loader/not-openapi.txt
