+++ test/fixtures/loader/openapi.yaml
