+++ test/fixtures/loader/refs/a.yaml
