+++ test/fixtures/loader/refs/definitions/b.yml
