+++ test/fixtures/loader/refs/openapi.yaml
