+++ test/fixtures/loader/refs/parameters.yaml
