+++ test/lib/server.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const server = require('../lib/server.js');
+const server = require('../../lib/server.js');
 
 describe('Server', () => {
     describe('loadHTML()', () => {
